backend/requirements.txt
fastapi
uvicorn
python-multipart
chromadb
langchain
langchain-community
langchain-openai
pypdf
python-dotenv
tiktoken
